requirements.txt
transformers>=4.30.0
datasets>=2.14.0
pymupdf>=1.22.5
python-dotenv>=1.0.0
groq>=0.5.0
torch>=2.0.0